refactor(accounting): extract GeneralLedgerReportService from AccountingService (Wave-4 P6)
Behavior-preserving final slice: getTrialBalance, getProfitLossFromJournal,
getBalanceSheetFromJournal, getCashFlowFromJournal, getEquityStatementFromJournal,
getGeneralLedger, getGeneralJournal + private sumAccountBalances/sumDebitInPeriod
moved verbatim into a new GeneralLedgerReportService; AccountingService delegates
(signatures + controller/ReportsService/MonthlyClose/Closing wiring unchanged).
Read-only journal-line reports, one cohesive service (intra-cluster calls kept).
Trial Balance (FINANCE/SHOP/ALL, isAllBalanced) verified via delegation.
