+++ apps/api/src/modules/accounting/accounting.module.ts
@@ -5,6 +5,7 @@ import { AccountingService } from './accounting.service';
 import { PeakExportService } from './peak-export.service';
 import { ReceivablesReportService } from './receivables-report.service';
 import { TransactionalReportService } from './transactional-report.service';
+import { GeneralLedgerReportService } from './general-ledger-report.service';
 import { BadDebtService } from './bad-debt.service';
 import { BadDebtProvisionCron } from './bad-debt-provision.cron';
 import { MonthlyCloseService } from './monthly-close.service';
@@ -24,6 +25,7 @@ import { PeakModule } from '../peak/peak.module';
     PeakExportService,
     ReceivablesReportService,
     TransactionalReportService,
+    GeneralLedgerReportService,
     BadDebtService,
     BadDebtProvisionCron,
     MonthlyCloseService,
