+++ apps/api/src/modules/accounting/accounting.service.spec.ts
@@ -4,6 +4,7 @@ import { AccountingService } from './accounting.service';
 import { PeakExportService } from './peak-export.service';
 import { ReceivablesReportService } from './receivables-report.service';
 import { TransactionalReportService } from './transactional-report.service';
+import { GeneralLedgerReportService } from './general-ledger-report.service';
 import { PrismaService } from '../../prisma/prisma.service';
 import { JournalAutoService } from '../journal/journal-auto.service';
 import { CompanyResolverService } from '../journal/company-resolver.service';
@@ -120,6 +121,7 @@ describe('AccountingService', () => {
         PeakExportService,
         ReceivablesReportService,
         TransactionalReportService,
+        GeneralLedgerReportService,
         { provide: PrismaService, useValue: prisma },
         { provide: JournalAutoService, useValue: journalAutoService },
         {
